+++ src/libtrx/include/libtrx/game/ui/elements/bar_lara_sprint.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool UI_LaraSprintBar(void);